+++ backend/src/navigator_handlers.py
