+++ backend/src/server.py
@@ -2,10 +2,10 @@ from flask import Flask, request, jsonify, send_file, redirect, Response
 from flask_restful import Resource, Api
 import requests
 import db
-import navigatorhandlers
-import landlordhandlers
-import propertyhandlers
-import authhandlers
+import navigator_handlers
+import landlord_handlers
+import property_handlers
+import auth_handlers
 import os
 import logging
 
@@ -42,16 +42,16 @@ supportedCrudEndpoints = [{
     "/navigator",
     "methods": [{
         "method": "GET",
-        "handler": navigatorhandlers.get_navigator
+        "handler": navigator_handlers.get_navigator
     }, {
         "method": "POST",
-        "handler": navigatorhandlers.post_navigator
+        "handler": navigator_handlers.post_navigator
     }, {
         "method": "PUT",
-        "handler": navigatorhandlers.put_navigator
+        "handler": navigator_handlers.put_navigator
     }, {
         "method": "DELETE",
-        "handler": navigatorhandlers.delete_navigator
+        "handler": navigator_handlers.delete_navigator
     }]
 }, {
     "name":
@@ -60,16 +60,16 @@ supportedCrudEndpoints = [{
     "/landlord",
     "methods": [{
         "method": "GET",
-        "handler": landlordhandlers.get_landlord
+        "handler": landlord_handlers.get_landlord
     }, {
         "method": "POST",
-        "handler": landlordhandlers.post_landlord
+        "handler": landlord_handlers.post_landlord
     }, {
         "method": "PUT",
-        "handler": landlordhandlers.put_landlord
+        "handler": landlord_handlers.put_landlord
     }, {
         "method": "DELETE",
-        "handler": landlordhandlers.delete_landlord
+        "handler": landlord_handlers.delete_landlord
     }]
 }, {
     "name":
@@ -78,16 +78,16 @@ supportedCrudEndpoints = [{
     "/property",
     "methods": [{
         "method": "GET",
-        "handler": propertyhandlers.get_property
+        "handler": property_handlers.get_property
     }, {
         "method": "POST",
-        "handler": propertyhandlers.post_property
+        "handler": property_handlers.post_property
     }, {
         "method": "PUT",
-        "handler": propertyhandlers.put_property
+        "handler": property_handlers.put_property
     }, {
         "method": "DELETE",
-        "handler": propertyhandlers.delete_property
+        "handler": property_handlers.delete_property
     }]
 }, {
     "name":
@@ -96,7 +96,7 @@ supportedCrudEndpoints = [{
     "/auth/register",
     "methods": [{
         "method": "POST",
-        "handler": authhandlers.register_new_user
+        "handler": auth_handlers.register_new_user
     }]
 }, {
     "name":
@@ -105,7 +105,7 @@ supportedCrudEndpoints = [{
     "/auth/login",
     "methods": [{
         "method": "POST",
-        "handler": authhandlers.login
+        "handler": auth_handlers.login
     }]
 }, {
     "name":
@@ -114,7 +114,7 @@ supportedCrudEndpoints = [{
     "/auth/status",
     "methods": [{
         "method": "GET",
-        "handler": authhandlers.get_login_status
+        "handler": auth_handlers.get_login_status
     }]
 }, {
     "name":
@@ -123,7 +123,7 @@ supportedCrudEndpoints = [{
     "/auth/logout",
     "methods": [{
         "method": "GET",
-        "handler": authhandlers.logout
+        "handler": auth_handlers.logout
     }]
 }]
